Fix for #155. Changed parseInt() to parseFloat() so the decimal places for the calculation are preserved and the error for the given use case is resolved. Unit test added and passing for the issue however the basic failure isn't generating with the correct code that the auditor is presenting. It's likely to be an issue with the unit test or unit test runner code rather than an actual failure to detect (works when run manually).

+++ Standards/WCAG2AAA/Sniffs/Principle1/Guideline1_4/1_4_3_Contrast.js
@@ -66,7 +66,7 @@ _global.HTMLCS_WCAG2AAA_Sniffs_Principle1_Guideline1_4_1_4_3_Contrast = {
                         // Calculate font size. Note that CSS 2.1 fixes a reference pixel
                         // as 96 dpi (hence "pixel ratio" workarounds for Hi-DPI devices)
                         // so this calculation should be safe.
-                        var fontSize      = parseInt(style.fontSize, 10) * (72 / 96);
+                        var fontSize      = parseFloat(style.fontSize, 10) * (72 / 96);
                         var minLargeSize  = 18;
 
                         if ((style.fontWeight === 'bold') || (parseInt(style.fontWeight, 10) >= 600)) {
